separate core(rdd) and streaming implicits to separate packages, so that when using core functions only it's not required to include spark-streaming dependency
