+++ src/main/scala/com/redislabs/provider/redis/streaming/package.scala
@@ -1,0 +1,5 @@
+package com.redislabs.provider.redis
+
+package object streaming extends RedisStreamingFunctions {
+
+}
